services/auth_service/pyproject.toml
[project]
name = "auth-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.121.2",
    "pydantic[email]>=2.12.4",
    "pydantic-settings>=2.12.0",
    "uvicorn>=0.38.0",
    "redis>=7.0.1",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.5.0",
    "httpx>=0.28.1",
]

[dependency-groups]
dev = [
    "deptry>=0.24.0",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.5",
]
